+++ modules/AWSPowerShell/Cmdlets/DataSync/Basic/New-DSYNLocationObjectStorage-Cmdlet.cs
@@ -108,8 +108,8 @@ namespace Amazon.PowerShell.Cmdlets.DSYN
         /// <summary>
         /// <para>
         /// <para>The name of the self-managed object storage server. This value is the IP address or
-        /// Domain Name Service (DNS) name of the object storage server. An agent uses this host
-        /// name to mount the object storage server in a network. </para>
+        /// Domain Name Service (DNS) name of the object storage server. An agent uses this hostname
+        /// to mount the object storage server in a network. </para>
         /// </para>
         /// </summary>
         #if !MODULAR