+++ modules/AWSPowerShell/Cmdlets/FSx/Basic/Restore-FSXVolumeFromSnapshot-Cmdlet.cs
@@ -54,9 +54,9 @@ namespace Amazon.PowerShell.Cmdlets.FSX
         /// <para>
         /// <para>The settings used when restoring the specified volume from snapshot. </para><ul><li><para><code>DELETE_INTERMEDIATE_SNAPSHOTS</code> - Deletes snapshots between the current
         /// state and the specified snapshot. If there are intermediate snapshots and this option
-        /// isn't used, <code>RestoreVolumeFromSnapshot</code> fails.</para></li><li><para><code>DELETE_CLONED_VOLUMES</code> - Deletes any volumes cloned from this volume.
-        /// If there are any cloned volumes and this option isn't used, <code>RestoreVolumeFromSnapshot</code>
-        /// fails.</para></li></ul>
+        /// isn't used, <code>RestoreVolumeFromSnapshot</code> fails.</para></li><li><para><code>DELETE_CLONED_VOLUMES</code> - Deletes any dependent clone volumes created
+        /// from intermediate snapshots. If there are any dependent clone volumes and this option
+        /// isn't used, <code>RestoreVolumeFromSnapshot</code> fails.</para></li></ul>
         /// </para>
         /// </summary>
         [System.Management.Automation.Parameter(ValueFromPipelineByPropertyName = true)]